+++ src/components/refRadarComponent/utils/index.js
@@ -9,7 +9,13 @@ function generateBaseCircles(numCircles, color) {
   return baseCircles;
 }
 
-export function BaseCircles({ svg, numCircles, colorCircles, radius, circleStroke }) {
+export function BaseCircles({
+  svg,
+  numCircles,
+  colorCircles,
+  radius,
+  circleStroke,
+}) {
   const baseCircles = generateBaseCircles(numCircles, colorCircles);
   const base = svg.append("g").attr("class", "base");
 
@@ -22,6 +28,24 @@ export function BaseCircles({ svg, numCircles, colorCircles, radius, circleStrok
       .attr("stroke", circle.color)
       .attr("stroke-width", circleStroke)
       .attr("fill", "none");
+  });
+}
+
+export function BaseLines({ lineAngles,radius,colorLines,strokeLines,opacityLines, svg }) {
+  // Dibujar líneas desde el centro hasta el radio especificado
+  lineAngles.forEach((angle) => {
+    const x2 = Math.cos(angle) * radius;
+    const y2 = Math.sin(angle) * (-radius + 1.5);
+
+    svg
+      .append("line")
+      .attr("x1", 0)
+      .attr("y1", 0)
+      .attr("x2", x2)
+      .attr("y2", y2)
+      .attr("stroke", colorLines)
+      .attr("stroke-width", strokeLines)
+      .attr("opacity", opacityLines);
   });
 }
